FCM notifications
New classes added and code refactored.
