+++ app/src/main/AndroidManifest.xml
@@ -47,6 +47,22 @@
 
         <service android:name=".LocationService"/>
 
+        <service
+            android:name="app.chaffer.MyFirebaseInstanceIDService">
+            <intent-filter>
+                <action
+                    android:name="com.google.firebase.INSTANCE_ID_EVENT"/>
+            </intent-filter>
+        </service>
+
+        <service
+            android:name="app.chaffer.GCMNotificationIntentService"
+            android:exported="false">
+            <intent-filter>
+                <action android:name="com.google.firebase.MESSAGING_EVENT" />
+            </intent-filter>
+        </service>
+
     </application>
 
 </manifest>
